+++ lib/sluggi.rb
@@ -1,5 +1,9 @@
 require "sluggi/version"
 require "sluggi/slug"
 require "sluggi/history"
+require "sluggi/validate_exclusion_of"
+require "sluggi/validate_presence"
+require "sluggi/validate_uniqueness"
+require "sluggi/model"
 require "sluggi/slugged"
 require "sluggi/generators/sluggi_generator"
