+++ lib/sluggi/validate_presence.rb
@@ -1,0 +1,9 @@
+module Sluggi
+  module ValidatePresence
+    extend ActiveSupport::Concern
+
+    included do
+      validates :slug, presence: true
+    end
+  end
+end
