+++ lib/sluggi/validate_uniqueness.rb
@@ -1,0 +1,9 @@
+module Sluggi
+  module ValidateUniqueness
+    extend ActiveSupport::Concern
+
+    included do
+      validates :slug, uniqueness: true
+    end
+  end
+end
